Move ompi_mpi_object_t to separate header file
