+++ ompi/request/mpi_object.h
@@ -1,0 +1,64 @@
+/* -*- Mode: C; c-basic-offset:4 ; indent-tabs-mode:nil -*- */
+/*
+ * Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
+ *                         University Research and Technology
+ *                         Corporation.  All rights reserved.
+ * Copyright (c) 2004-2022 The University of Tennessee and The University
+ *                         of Tennessee Research Foundation.  All rights
+ *                         reserved.
+ * Copyright (c) 2004-2005 High Performance Computing Center Stuttgart,
+ *                         University of Stuttgart.  All rights reserved.
+ * Copyright (c) 2004-2005 The Regents of the University of California.
+ *                         All rights reserved.
+ * Copyright (c) 2006-2017 Cisco Systems, Inc.  All rights reserved
+ * Copyright (c) 2009-2012 Oracle and/or its affiliates.  All rights reserved.
+ * Copyright (c) 2012      Oak Ridge National Labs.  All rights reserved.
+ * Copyright (c) 2015-2017 Los Alamos National Security, LLC.  All rights
+ *                         reserved.
+ * Copyright (c) 2018      FUJITSU LIMITED.  All rights reserved.
+ * $COPYRIGHT$
+ *
+ * Additional copyrights may follow
+ *
+ * $HEADER$
+ */
+/**
+ * @file
+ * Union holding pointers to communicators, windows, and files
+ *
+ */
+
+#ifndef OMPI_MPI_OBJECT_H
+#define OMPI_MPI_OBJECT_H
+
+#include "ompi_config.h"
+
+BEGIN_C_DECLS
+
+/**
+ * Forward declaration
+ */
+struct ompi_communicator_t;
+
+/**
+ * Forward declaration
+ */
+struct ompi_win_t;
+
+/**
+ * Forward declaration
+ */
+struct ompi_file_t;
+
+/**
+ * Union for holding several different MPI pointer types on the request
+ */
+typedef union ompi_mpi_object_t {
+    struct ompi_communicator_t *comm;
+    struct ompi_file_t *file;
+    struct ompi_win_t *win;
+} ompi_mpi_object_t;
+
+END_C_DECLS
+
+#endif // OMPI_MPI_OBJECT_H
